requirements.txt
# Core dependencies
opencv-python>=4.5.0
pyzbar>=0.1.8
Pillow>=8.0.0
numpy>=1.20.0

# Google Sheets integration
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0

# GUI and automation
pyautogui>=0.9.50

# Data validation
email-validator>=1.1.0

# Development and testing (optional)
pytest>=6.0.0
pytest-cov>=2.10.0 